+++ Tests/EditorLogicTests/EditorLogicTests.cpp
@@ -167,6 +167,7 @@ int main(int argc, char** argv)
 		testSubwooferRoutingReadsRoundTripEverySetter();
 		testSubwooferRoutingSourceLfeGainReaders();
 		testSubwooferRoutingPreviewRateAndTrim();
+		testSubwooferRoutingPresetDisplayName();
 		testAnalysisRequestFence();
 		testImpulseMeasurementFindsTheDelay();
 		testImpulseMeasurementAtTheBlockStart();